Separate the logic between resolver and validator
furthermore, it's now possible to resolve/validate the trust chain
starting from the provided RP's EC or trust chain (this will be useful
for explicit client registration), as well as by the entity_id
in the standard phase of federation entity discovery
